+++ app/_components/vision.tsx
@@ -1,64 +1,24 @@
-'use client'
-
-import React, { useState } from 'react';
-import { Accordion, AccordionHeader, AccordionBody } from './accordionWrapper'
-
-interface IconProps {
-  id: number;
-  open: number;
-}
-
-const Icon: React.FC<IconProps> = ({ id, open }) => {
-  return (
-    <svg
-      xmlns="(link unavailable)"
-      fill="none"
-      viewBox="0 0 24 24"
-      strokeWidth={2}
-      stroke="currentColor"
-      className={`${id === open ? 'rotate-180' : ''} h-5 w-5 transition-transform`}
-    >
-      <path strokeLinecap="round" strokeLinejoin="round" d="M19.5 8.25l-7.5 7.5-7.5-7.5" />
-    </svg>
-  );
-};
+import visionImage from '../../public/images/vision.png'
+import Image from 'next/image';
+import Accordions from './accordions';
 
 const Vision = () => {
-  const [open, setOpen] = useState(0);
-
-  const handleOpen = (value: number) => setOpen(open === value ? 0 : value);
-
   return (
-    <>
-      <Accordion open={open === 1} icon={<Icon id={1} open={open} />} onClick={() => handleOpen(1)} placeholder={undefined} onPointerEnterCapture={undefined} onPointerLeaveCapture={undefined}>
-        <AccordionHeader placeholder={undefined} onPointerEnterCapture={undefined} onPointerLeaveCapture={undefined}>What is Material Tailwind?</AccordionHeader>
-        <AccordionBody>
-          We're not always in the position that we want to be at. We're constantly
-          growing. We're constantly making mistakes. We're constantly trying to express
-          ourselves and actualize our dreams.
-        </AccordionBody>
-      </Accordion>
-      <Accordion open={open === 2} icon={<Icon id={2} open={open} />} onClick={() => handleOpen(2)} placeholder={undefined} onPointerEnterCapture={undefined} onPointerLeaveCapture={undefined}>
-        <AccordionHeader placeholder={undefined} onPointerEnterCapture={undefined} onPointerLeaveCapture={undefined}>
-          How to use Material Tailwind?
-        </AccordionHeader>
-        <AccordionBody>
-          We're not always in the position that we want to be at. We're constantly
-          growing. We're constantly making mistakes. We're constantly trying to express
-          ourselves and actualize our dreams.
-        </AccordionBody>
-      </Accordion>
-      <Accordion open={open === 3} icon={<Icon id={3} open={open} />} onClick={() => handleOpen(3)}  placeholder={undefined} onPointerEnterCapture={undefined} onPointerLeaveCapture={undefined}>
-        <AccordionHeader  placeholder={undefined} onPointerEnterCapture={undefined} onPointerLeaveCapture={undefined}>
-          What can I do with Material Tailwind?
-        </AccordionHeader>
-        <AccordionBody>
-          We're not always in the position that we want to be at. We're constantly
-          growing. We're constantly making mistakes. We're constantly trying to express
-          ourselves and actualize our dreams.
-        </AccordionBody>
-      </Accordion>
-    </>
+    <section className="flex h-screen">
+      <div className='site_container flex justify-between '>
+        <div className="w-1/2">
+          <div className="flex flex-col justify-center h-auto ">
+          <h2 className="text-white uppercase text-[38px]  font-bold">our vision is to support the innovation of AI blockchain <br/> while prioritizing communities and democratizing profits</h2>            <Image
+              src={visionImage}
+              alt="Image description"
+            />
+          </div>
+        </div>
+        <div className="w-1/2 h-full">
+          <Accordions />
+        </div>
+      </div>
+    </section>
   );
 };
