+++ next.config.mjs
@@ -1,6 +1,5 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-    videos: ['public/videos/**/*']
 };
 
 export default nextConfig;
